Refactor invoice form and add submit functionality
